+++ src/main/java/edu/byu/cs/controller/CasController.java
@@ -1,6 +1,8 @@
 package edu.byu.cs.controller;
 
 import edu.byu.cs.canvas.CanvasException;
+import edu.byu.cs.controller.exception.BadRequestException;
+import edu.byu.cs.controller.exception.InternalServerException;
 import edu.byu.cs.dataAccess.DataAccessException;
 import edu.byu.cs.model.User;
 import edu.byu.cs.properties.ApplicationProperties;
