+++ src/main/java/edu/byu/cs/controller/exception/BadRequestException.java
@@ -1,4 +1,4 @@
-package edu.byu.cs.controller;
+package edu.byu.cs.controller.exception;
 
 public class BadRequestException extends Exception {
     public BadRequestException(String message, Throwable cause) {
