+++ src/main/java/edu/byu/cs/controller/exception/PriorRepoClaimBlockageException.java
@@ -1,4 +1,4 @@
-package edu.byu.cs.controller;
+package edu.byu.cs.controller.exception;
 
 public class PriorRepoClaimBlockageException extends Exception {
     private static final String secrets = "btw im a teapot";
