+++ src/main/java/edu/byu/cs/service/UserService.java
@@ -1,8 +1,8 @@
 package edu.byu.cs.service;
 
-import edu.byu.cs.controller.BadRequestException;
-import edu.byu.cs.controller.InternalServerException;
-import edu.byu.cs.controller.PriorRepoClaimBlockageException;
+import edu.byu.cs.controller.exception.BadRequestException;
+import edu.byu.cs.controller.exception.InternalServerException;
+import edu.byu.cs.controller.exception.PriorRepoClaimBlockageException;
 import edu.byu.cs.dataAccess.DaoService;
 import edu.byu.cs.dataAccess.DataAccessException;
 import edu.byu.cs.model.RepoUpdate;
@@ -109,7 +109,8 @@ public class UserService {
     /**
      * Checks to see if anyone currently or previously (other than the provided user) has claimed the provided repoUrl.
      * returns if the repo is available. it will throw otherwise, containing a message why
-     * @param url the repoUrl to check if currently or previously claimed
+     *
+     * @param url   the repoUrl to check if currently or previously claimed
      * @param netId the user trying to claim the url, so that they can claim urls they previously claimed
      * @return null if the repo is available for that user. returns the update that prevents the user from claiming the url.
      */
